Remove jslib authService and unused loginStrategies
