+++ src/commands/login.command.ts
@@ -1,11 +1,11 @@
 import * as inquirer from "inquirer";
 
-import { AuthService } from "@/jslib/common/src/abstractions/auth.service";
 import { ApiLogInCredentials } from "@/jslib/common/src/models/domain/logInCredentials";
 import { Response } from "@/jslib/node/src/cli/models/response";
 import { MessageResponse } from "@/jslib/node/src/cli/models/response/messageResponse";
 
 import { Utils } from "../../jslib/common/src/misc/utils";
+import { AuthService } from "../abstractions/auth.service";
 
 export class LoginCommand {
   private canInteract: boolean;
